Refactor tests to their own directory and individualized per API
